refactors GM page, extracting table settings
